requirements.txt
flet[all]
python-dotenv
dataclasses-json
plotly
kaleido
flet-charts